--- Packages/io.chainsafe.web3-unity/Runtime/Scripts/Samples/IpfsSample.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: d61a1e7c98424af8b0ed86a6516bbcc0
-timeCreated: 1692280714
