+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.5.0/Web3.Unity Samples/Scripts/Scenes/SampleMain/IPFS.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b70d421e7223b8049aaf18e07497da9b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
