+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.5.0/Web3.Unity Samples/Scripts/Scenes/SampleMain/IPFS/IPFSCalls.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9a86b695c20675b43ba7a0f0a496749c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
